requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
pymssql==2.2.11
openpyxl==3.1.2
weasyprint==60.2
lxml==5.1.0
cryptography==41.0.7
python-dotenv==1.0.0
